// src/app/api-client/model/licensingSummary.ts
/**
 * Horde Model Reference API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * Represents aggregate counts useful to policy consumers and user interfaces.
 */
export interface LicensingSummary {
  total_assets: number;
  commercial_use: { [key: string]: number };
  redistribution: { [key: string]: number };
  licenses: { [key: string]: number };
}